src/torch_web/requirements.txt
# Main Environment
fastapi
python-socketio
sqlmodel
uvicorn[standard]
python-dotenv
requests
psycopg2

# Task Dependencies
imagehash
minio
Pillow
azure-storage-blob
azure-identity
azure-cognitiveservices-vision-computervision
azure-keyvault-secrets
paramiko
